requirements.txt
python-telegram-bot==21.6
aiosqlite==0.19.0

